main: upcrease reported_cfg_size
cfgfs_read() seems to be always called with buffers 4096 bytes big. this way
it'll use all of it instead of just the first 1024 bytes

the size may depend on glibc version. if an older version has different behavior
then it shall be fixed once found out

an alternative way could be to put the size in the buffer_list_item struct or
something